+++ pom.xml
@@ -52,16 +52,6 @@
 			<artifactId>jmeter-plugins-casutg</artifactId>
 			<version>2.9</version>
 		</dependency>
-		<dependency>
-			<groupId>org.apache.jmeter</groupId>
-			<artifactId>ApacheJMeter_monitors</artifactId>
-			<version>3.1</version>
-		</dependency>
-		<dependency>
-			<groupId>kg.apc</groupId>
-			<artifactId>jmeter-plugins-autostop</artifactId>
-			<version>0.1</version>
-		</dependency>
 	</dependencies>
 
 	<build>
