+++ src/main/java/com/jmeterpractice/practice/controllers/MainController.java
@@ -1,6 +1,6 @@
 package com.jmeterpractice.practice.controllers;
 
-import com.jmeterpractice.practice.helpclasses.JmeterTesting;
+import com.jmeterpractice.practice.modules.JmeterTesting;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
