--- src/test/java/com/jmeterpractice/practice/helpclasses/JmeterTestingTest.java
@@ -1,12 +1,0 @@
-package com.jmeterpractice.practice.helpclasses;
-
-import org.junit.Test;
-
-public class JmeterTestingTest {
-
-    @Test
-    public void start() {
-        JmeterTesting.getJmeterTest().start();
-    }
-
-}
