+++ src/test/java/com/jmeterpractice/practice/modules/JmeterTestingTest.java
@@ -1,0 +1,12 @@
+package com.jmeterpractice.practice.modules;
+
+import org.junit.Test;
+
+public class JmeterTestingTest {
+
+    @Test
+    public void start() {
+        JmeterTesting.getJmeterTest().start();
+    }
+
+}
